Test adjacent block caret
